moving mantle inside aws service
